Zemax Updated Optical Design Run1532
This is the most recent updated 2019-02-15 optical design of ROSIE IFU. It has 21 slits for both the inner and outer pseudoslits, individual slitlet width 0.6" and length 53.5". Its field of view is 53.5" x 50.4", with some gaps between the 50.4" direction. This design fixes the problem of overfilling the 2nd achromats and spacing them too close together in the outer pseudoslits.  I have checked for height collisions between the fold mirror 3s and the beam after the achromats in both the inner and outer pseudoslits.  Last time I checked, the widest fanned configurations in the outer pseudoslit don't run in to the enclosure, but only Leon will be able to tell for sure.